+++ .flake8
@@ -11,5 +11,4 @@ per-file-ignores =
     tests/test_forward_references.py:E800
     tests/schema/test_resolvers.py:E800
     tests/types/test_string_annotations.py:E800
-    tests/federation/test_printer.py:E800
-    tests/federation/test_printer.py:E501
+    tests/federation/printer/*:E501
