+++ tests/federation/printer/test_requires.py
@@ -1,0 +1,85 @@
+# type: ignore
+
+import textwrap
+from typing import List
+
+import strawberry
+
+
+def test_fields_requires_are_printed_correctly():
+    global Review
+
+    @strawberry.federation.type
+    class User:
+        username: str
+
+    @strawberry.federation.type(keys=["upc"], extend=True)
+    class Product:
+        upc: str = strawberry.federation.field(external=True)
+        field1: str = strawberry.federation.field(external=True)
+        field2: str = strawberry.federation.field(external=True)
+        field3: str = strawberry.federation.field(external=True)
+
+        @strawberry.federation.field(requires=["field1", "field2", "field3"])
+        def reviews(self) -> List["Review"]:
+            return []
+
+    @strawberry.federation.type
+    class Review:
+        body: str
+        author: User
+        product: Product
+
+    @strawberry.federation.type
+    class Query:
+        @strawberry.field
+        def top_products(self, first: int) -> List[Product]:
+            return []
+
+    schema = strawberry.federation.Schema(query=Query)
+
+    expected = """
+        directive @external on FIELD_DEFINITION
+
+        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
+
+        directive @requires(fields: _FieldSet!) on FIELD_DEFINITION
+
+        extend type Product @key(fields: "upc") {
+          upc: String! @external
+          field1: String! @external
+          field2: String! @external
+          field3: String! @external
+          reviews: [Review!]! @requires(fields: "field1 field2 field3")
+        }
+
+        type Query {
+          _service: _Service!
+          _entities(representations: [_Any!]!): [_Entity]!
+          topProducts(first: Int!): [Product!]!
+        }
+
+        type Review {
+          body: String!
+          author: User!
+          product: Product!
+        }
+
+        type User {
+          username: String!
+        }
+
+        scalar _Any
+
+        union _Entity = Product
+
+        scalar _FieldSet
+
+        type _Service {
+          sdl: String!
+        }
+    """
+
+    assert schema.as_str() == textwrap.dedent(expected).strip()
+
+    del Review
