--- tests/federation/test_printer.py
@@ -1,730 +1,0 @@
-# type: ignore
-
-import textwrap
-from typing import List
-
-import strawberry
-from strawberry.federation.schema_directives import Key
-from strawberry.schema.config import StrawberryConfig
-from strawberry.schema_directive import Location
-
-
-def test_entities_type_when_no_type_has_keys():
-    global Review
-
-    @strawberry.federation.type
-    class User:
-        username: str
-
-    @strawberry.federation.type(keys=["upc"], extend=True)
-    class Product:
-        upc: str = strawberry.federation.field(external=True)
-        reviews: List["Review"]
-
-    @strawberry.federation.type
-    class Review:
-        body: str
-        author: User
-        product: Product
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(query=Query)
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        extend type Product @key(fields: "upc") {
-          upc: String! @external
-          reviews: [Review!]!
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        type Review {
-          body: String!
-          author: User!
-          product: Product!
-        }
-
-        type User {
-          username: String!
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-    del Review
-
-
-def test_entities_extending_interface():
-    @strawberry.interface
-    class SomeInterface:
-        id: strawberry.ID
-
-    @strawberry.federation.type(keys=["upc"], extend=True)
-    class Product(SomeInterface):
-        upc: str = strawberry.federation.field(external=True)
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(query=Query)
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        extend type Product implements SomeInterface @key(fields: "upc") {
-          id: ID!
-          upc: String! @external
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        interface SomeInterface {
-          id: ID!
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-
-def test_fields_requires_are_printed_correctly():
-    global Review
-
-    @strawberry.federation.type
-    class User:
-        username: str
-
-    @strawberry.federation.type(keys=["upc"], extend=True)
-    class Product:
-        upc: str = strawberry.federation.field(external=True)
-        field1: str = strawberry.federation.field(external=True)
-        field2: str = strawberry.federation.field(external=True)
-        field3: str = strawberry.federation.field(external=True)
-
-        @strawberry.federation.field(requires=["field1", "field2", "field3"])
-        def reviews(self) -> List["Review"]:
-            return []
-
-    @strawberry.federation.type
-    class Review:
-        body: str
-        author: User
-        product: Product
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(query=Query)
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        directive @requires(fields: _FieldSet!) on FIELD_DEFINITION
-
-        extend type Product @key(fields: "upc") {
-          upc: String! @external
-          field1: String! @external
-          field2: String! @external
-          field3: String! @external
-          reviews: [Review!]! @requires(fields: "field1 field2 field3")
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        type Review {
-          body: String!
-          author: User!
-          product: Product!
-        }
-
-        type User {
-          username: String!
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-    del Review
-
-
-def test_field_provides_are_printed_correctly_camel_case_on():
-    global Review
-
-    @strawberry.federation.type
-    class User:
-        username: str
-
-    @strawberry.federation.type(keys=["upc"], extend=True)
-    class Product:
-        upc: str = strawberry.federation.field(external=True)
-        the_name: str = strawberry.federation.field(external=True)
-        reviews: List["Review"]
-
-    @strawberry.federation.type
-    class Review:
-        body: str
-        author: User
-        product: Product = strawberry.federation.field(provides=["name"])
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(
-        query=Query, config=StrawberryConfig(auto_camel_case=True)
-    )
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        directive @provides(fields: _FieldSet!) on FIELD_DEFINITION
-
-        extend type Product @key(fields: "upc") {
-          upc: String! @external
-          theName: String! @external
-          reviews: [Review!]!
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        type Review {
-          body: String!
-          author: User!
-          product: Product! @provides(fields: "name")
-        }
-
-        type User {
-          username: String!
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-    del Review
-
-
-def test_field_provides_are_printed_correctly_camel_case_off():
-    global Review
-
-    @strawberry.federation.type
-    class User:
-        username: str
-
-    @strawberry.federation.type(keys=["upc"], extend=True)
-    class Product:
-        upc: str = strawberry.federation.field(external=True)
-        the_name: str = strawberry.federation.field(external=True)
-        reviews: List["Review"]
-
-    @strawberry.federation.type
-    class Review:
-        body: str
-        author: User
-        product: Product = strawberry.federation.field(provides=["name"])
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(
-        query=Query, config=StrawberryConfig(auto_camel_case=False)
-    )
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        directive @provides(fields: _FieldSet!) on FIELD_DEFINITION
-
-        extend type Product @key(fields: "upc") {
-          upc: String! @external
-          the_name: String! @external
-          reviews: [Review!]!
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          top_products(first: Int!): [Product!]!
-        }
-
-        type Review {
-          body: String!
-          author: User!
-          product: Product! @provides(fields: "name")
-        }
-
-        type User {
-          username: String!
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-    del Review
-
-
-def test_multiple_keys():
-    # also confirm that the "resolvable: True" works
-    global Review
-
-    @strawberry.federation.type
-    class User:
-        username: str
-
-    @strawberry.federation.type(keys=[Key("upc", True)], extend=True)
-    class Product:
-        upc: str = strawberry.federation.field(external=True)
-        reviews: List["Review"]
-
-    @strawberry.federation.type(keys=["body"])
-    class Review:
-        body: str
-        author: User
-        product: Product
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(query=Query)
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        extend type Product @key(fields: "upc", resolvable: true) {
-          upc: String! @external
-          reviews: [Review!]!
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        type Review @key(fields: "body") {
-          body: String!
-          author: User!
-          product: Product!
-        }
-
-        type User {
-          username: String!
-        }
-
-        scalar _Any
-
-        union _Entity = Product | Review
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-    del Review
-
-
-def test_field_shareable_printed_correctly():
-    @strawberry.interface
-    class SomeInterface:
-        id: strawberry.ID
-
-    @strawberry.federation.type(keys=["upc"], extend=True, shareable=True)
-    class Product(SomeInterface):
-        upc: str = strawberry.federation.field(external=True, shareable=True)
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(query=Query)
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        directive @shareable on FIELD_DEFINITION | OBJECT
-
-        extend type Product implements SomeInterface @key(fields: "upc") @shareable {
-          id: ID!
-          upc: String! @external @shareable
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        interface SomeInterface {
-          id: ID!
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-
-def test_field_tag_printed_correctly():
-    @strawberry.interface
-    class SomeInterface:
-        id: strawberry.ID
-
-    @strawberry.federation.type(keys=["upc"], extend=True)
-    class Product(SomeInterface):
-        upc: str = strawberry.federation.field(external=True, tags=["myTag"])
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(query=Query)
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        directive @tag(name: String!) on FIELD_DEFINITION | INTERFACE | OBJECT | UNION | ARGUMENT_DEFINITION | SCALAR | ENUM | ENUM_VALUE | INPUT_OBJECT | INPUT_FIELD_DEFINITION
-
-        extend type Product implements SomeInterface @key(fields: "upc") {
-          id: ID!
-          upc: String! @external @tag(name: "myTag")
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        interface SomeInterface {
-          id: ID!
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-
-def test_field_override_printed_correctly():
-    @strawberry.interface
-    class SomeInterface:
-        id: strawberry.ID
-
-    @strawberry.federation.type(keys=["upc"], extend=True)
-    class Product(SomeInterface):
-        upc: str = strawberry.federation.field(external=True, override="mySubGraph")
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(query=Query)
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        directive @override(from: String!) on FIELD_DEFINITION
-
-        extend type Product implements SomeInterface @key(fields: "upc") {
-          id: ID!
-          upc: String! @external @override(from: "mySubGraph")
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        interface SomeInterface {
-          id: ID!
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-
-def test_field_inaccessible_printed_correctly():
-    @strawberry.federation.interface(inaccessible=True)
-    class AnInterface:
-        id: strawberry.ID
-
-    @strawberry.interface
-    class SomeInterface:
-        id: strawberry.ID
-        a_field: str = strawberry.federation.field(inaccessible=True)
-
-    @strawberry.federation.type(keys=["upc"], extend=True)
-    class Product(SomeInterface):
-        upc: str = strawberry.federation.field(external=True, inaccessible=True)
-
-    @strawberry.federation.type
-    class Query:
-        @strawberry.field
-        def top_products(self, first: int) -> List[Product]:
-            return []
-
-    schema = strawberry.federation.Schema(query=Query, types=[AnInterface])
-
-    expected = """
-        directive @external on FIELD_DEFINITION
-
-        directive @inaccessible on FIELD_DEFINITION | OBJECT | INTERFACE | UNION | ARGUMENT_DEFINITION | SCALAR | ENUM | ENUM_VALUE | INPUT_OBJECT | INPUT_FIELD_DEFINITION
-
-        directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-        interface AnInterface @inaccessible {
-          id: ID!
-        }
-
-        extend type Product implements SomeInterface @key(fields: "upc") {
-          id: ID!
-          aField: String! @inaccessible
-          upc: String! @external @inaccessible
-        }
-
-        type Query {
-          _service: _Service!
-          _entities(representations: [_Any!]!): [_Entity]!
-          topProducts(first: Int!): [Product!]!
-        }
-
-        interface SomeInterface {
-          id: ID!
-          aField: String! @inaccessible
-        }
-
-        scalar _Any
-
-        union _Entity = Product
-
-        scalar _FieldSet
-
-        type _Service {
-          sdl: String!
-        }
-    """
-
-    assert schema.as_str() == textwrap.dedent(expected).strip()
-
-
-def test_additional_schema_directives_printed_correctly_object():
-    @strawberry.schema_directive(locations=[Location.OBJECT])
-    class CacheControl:
-        max_age: int
-
-    @strawberry.federation.type(
-        keys=["id"], shareable=True, extend=True, directives=[CacheControl(max_age=42)]
-    )
-    class FederatedType:
-        id: strawberry.ID
-
-    @strawberry.type
-    class Query:
-        federatedType: FederatedType
-
-    expected_type = """
-    directive @CacheControl(max_age: Int!) on OBJECT
-
-    directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-    directive @shareable on FIELD_DEFINITION | OBJECT
-
-    extend type FederatedType @key(fields: "id") @shareable @CacheControl(max_age: 42) {
-      id: ID!
-    }
-
-    type Query {
-      federatedType: FederatedType!
-    }
-    """
-
-    schema = strawberry.Schema(
-        query=Query, config=StrawberryConfig(auto_camel_case=False)
-    )
-    assert schema.as_str() == textwrap.dedent(expected_type).strip()
-
-
-def test_additional_schema_directives_printed_in_order_object():
-    @strawberry.schema_directive(locations=[Location.OBJECT])
-    class CacheControl0:
-        max_age: int
-
-    @strawberry.schema_directive(locations=[Location.OBJECT])
-    class CacheControl1:
-        min_age: int
-
-    @strawberry.federation.type(
-        keys=["id"],
-        shareable=True,
-        extend=True,
-        directives=[CacheControl0(max_age=42), CacheControl1(min_age=42)],
-    )
-    class FederatedType:
-        id: strawberry.ID
-
-    @strawberry.type
-    class Query:
-        federatedType: FederatedType
-
-    expected_type = """
-    directive @CacheControl0(max_age: Int!) on OBJECT
-
-    directive @CacheControl1(min_age: Int!) on OBJECT
-
-    directive @key(fields: _FieldSet!, resolvable: Boolean = true) on OBJECT | INTERFACE
-
-    directive @shareable on FIELD_DEFINITION | OBJECT
-
-    extend type FederatedType @key(fields: "id") @shareable @CacheControl0(max_age: 42) @CacheControl1(min_age: 42) {
-      id: ID!
-    }
-
-    type Query {
-      federatedType: FederatedType!
-    }
-    """
-
-    schema = strawberry.Schema(
-        query=Query, config=StrawberryConfig(auto_camel_case=False)
-    )
-    assert schema.as_str() == textwrap.dedent(expected_type).strip()
